SVMMessage::{get_durable_nonce, get_ix_signers} (#3287)
