+++ runtime/src/lib.rs
@@ -17,7 +17,6 @@ pub mod genesis_utils;
 pub mod installed_scheduler_pool;
 pub mod loader_utils;
 pub mod non_circulating_supply;
-mod nonce_extraction;
 pub mod prioritization_fee;
 pub mod prioritization_fee_cache;
 pub mod rent_collector;
