+++ svm-transaction/src/lib.rs
@@ -2,3 +2,5 @@ pub mod instruction;
 pub mod message_address_table_lookup;
 pub mod svm_message;
 pub mod svm_transaction;
+
+mod tests;
